day 11

+++ Arcesium Interview Question/xor.cpp
@@ -4,19 +4,20 @@ using namespace std;
 int main(){
 	int n;
 	cin>>n;
-	long number=0;
+	long long number=0;
 	vector<int> b;
 	b.push_back(1);
 	for(long i=1;i<n;i++){
 		b.push_back(0);
 	}
-	int base = 1;
-	for(long i=b.size()-1;i>=0;i--){
+	long long base = 1;
+	for(long long i=b.size()-1;i>=0;i--){
 		number += b[i]*base;
 		base = base*2;
 	}
-	long prev = number;
-	for(int i=0;i<n-1;i++){
+	long long prev = number;
+	for(long long i=0;i<n-1;i++){
+//		cout<<number<<" "<<prev;
 		prev=prev>>1;
 		number=number^prev;
 		prev=number;
@@ -29,7 +30,7 @@ int main(){
 	}
 	base=1;
 	number=0;
-	for(long i=0;i<v.size();i++){
+	for(long long i=0;i<v.size();i++){
 		number+=v[i]*base;
 		base=base*2;
 	}
